requirements.txt
flask
flask-cors
joblib
pandas
scikit-learn
mysql-connector-python
azure-ai-anomalydetector
azure-core 
RandomForestClassifier
